Refactorizacion de codigo y alertas agregadas
